fin by rate
Додано кількість товару, її оновленя, видалення товару за 'немає у наявності'. перегляд усіх замовлень через сторінку адміну, виправлені відступи, перенесено відображеня замовлення до його картки

+++ server/controllers/orderConroller.js
@@ -11,18 +11,24 @@ class OrderController {
             postcode: req.body.postcode,
             addressee: req.body.addressee
         }
-        const order = await Order.create(newOrder)
 
-        const basket = await BasketDevice.findAll({ where: {basketId: req.body.id}})
-        basket.forEach(i =>
-            OrderDevice.create({
-                orderId: order.id,
-                deviceId: i.deviceId,
-                basketId: i.id,
-            }),
-        await BasketDevice.destroy({where: {basketId: req.body.id}})
+        const basket = await BasketDevice.findAll({where: {basketId: req.body.id}})
+
+        if (basket.length >= 1)
+        {
+            const order = await Order.create(newOrder)
+                basket.forEach(i =>
+                OrderDevice.create({
+                    orderId: order.id,
+                    deviceId: i.deviceId,
+                    basketId: i.id,
+                }),
+            await BasketDevice.destroy({where: {basketId: req.body.id}})
         )
-        res.status(201).json(order)
+            res.status(201).json(order)
+    }
+        res.status(404)
+        console.log("haven't devices")
     }
 
     async getAll(req,res){
@@ -34,7 +40,6 @@ class OrderController {
         const {id} = req.params
         const date = await Order.findAll({where: {userId: id}} )
           // delete the dot and everything after
-
         return res.json(date)
     }
     async getUserOrderList(req,res){
@@ -44,6 +49,14 @@ class OrderController {
                 model: Device
             }, where: {orderId: id}});
         return res.json(a)
+    }
+    async updateUserOrder(req,res){
+        const {_id,_status} = req.params
+        const device = await Order.update(
+            {status: _status},
+            {where: {id: _id}}
+        )
+        return res.json(device)
     }
 }
